renamed movieList to moviesList
