+++ src/App.js
@@ -1,11 +1,11 @@
 import { Navigation } from "./common/Navigation";
 import { PageList } from "./common/PageList";
-import { MovieList } from "./features/MovieList";
+import { MoviesList } from "./features/MoviesList";
 
 export const App = () => (
   <>
     <Navigation />
-    <MovieList />
+    <MoviesList />
     <PageList />
   </>
 );
