+++ src/rootSaga.js
@@ -1,5 +1,5 @@
 import { all } from "@redux-saga/core/effects";
-import { moviesSaga } from "./features/MovieList/moviesSaga";
+import { moviesSaga } from "./features/MoviesList/moviesSaga";
 
 export default function* rootSaga() {
   yield all([moviesSaga()]);
