+++ src/store.js
@@ -1,6 +1,6 @@
 import createSagaMiddleware from "@redux-saga/core";
 import { configureStore, combineReducers } from "@reduxjs/toolkit";
-import moviesReducer from "./features/MovieList/moviesSlice";
+import moviesReducer from "./features/MoviesList/moviesSlice";
 import rootSaga from "./rootSaga";
 
 const sagaMiddleware = createSagaMiddleware();
